+++ emBODY/eBcode/arch-arm/board/amcbldc/application/proj/amcbldc-application2.uvprojx
@@ -10,8 +10,7 @@
       <TargetName>amcbldc-app-motorhal2</TargetName>
       <ToolsetNumber>0x4</ToolsetNumber>
       <ToolsetName>ARM-ADS</ToolsetName>
-      <pArmCC>6190000::V6.19::.\ARMCLANG_6_19</pArmCC>
-      <pCCUsed>6190000::V6.19::.\ARMCLANG_6_19</pCCUsed>
+      <pCCUsed>6220000::V6.22::ARMCLANG</pCCUsed>
       <uAC6>1</uAC6>
       <TargetOption>
         <TargetCommonOption>
@@ -315,7 +314,7 @@
           </ArmAdsMisc>
           <Cads>
             <interw>1</interw>
-            <Optim>6</Optim>
+            <Optim>1</Optim>
             <oTime>0</oTime>
             <SplitLS>0</SplitLS>
             <OneElfS>1</OneElfS>
@@ -799,6 +798,11 @@
               <FileName>embot_hw_bsp_amcbldc.cpp</FileName>
               <FileType>8</FileType>
               <FilePath>..\..\bsp\embot_hw_bsp_amcbldc.cpp</FilePath>
+            </File>
+            <File>
+              <FileName>embot_hw_analog_bsp_amcbldc.cpp</FileName>
+              <FileType>8</FileType>
+              <FilePath>.\embot_hw_analog_bsp_amcbldc.cpp</FilePath>
             </File>
           </Files>
         </Group>
@@ -2111,6 +2115,11 @@
               <FileName>embot_hw_bsp_amcbldc.cpp</FileName>
               <FileType>8</FileType>
               <FilePath>..\..\bsp\embot_hw_bsp_amcbldc.cpp</FilePath>
+            </File>
+            <File>
+              <FileName>embot_hw_analog_bsp_amcbldc.cpp</FileName>
+              <FileType>8</FileType>
+              <FilePath>.\embot_hw_analog_bsp_amcbldc.cpp</FilePath>
             </File>
           </Files>
         </Group>
